Move action getDataEatingPlace to another file and changing the way it is called
